Fix log filter (#62)
* Update the version of rollbar-java 1.0.0 -> 1.0.1

* RegexFilter only applies to the log message not the entire log string

I was hoping I could apply the filter to the namespace, but looking at
the implementation of the Filter code, it doesn't do that instead it
just looks at the body part of the message. I could apply my own filter,
but the messages I'm trying to squanch have a very specific form.