Implement connect and migrate for EmbeddedDatabase
The original init closure was split between those two methods and now we
can control the different stages of the underlying database connection.

In addition, the way this class is instantiated is now different. Before
we only expected the caller to send the directory location, but now
we're expecting a JDBC string.

Since both were Strings, no signature refactor is needed, but we need to
update the call site.