Move databaseModule to Database.module
Doing this to remove the floating function `databaseModule` and
associate the module with the `Database' interface.

The other difference is that we're passing `isEmbedded` instead of
making that determination for the caller. This still could be cleaned up
a bit more if we were more specific about which `Database` instance
we're looking for.