Use MessageChannel instead of textChannel
This is a massive change, all channels now allow text to be used,
therefore, relying on `defaultChannel` is less of a need now.

This change also allows us to send commands from voice channel chats. By
default, the `defaultChannel` will be respected (not exhaustive, I'm
sure I missed a spot) otherwise the channel where the command was issued
will be used OR that same voice channel since we can now write to it.

This originally was changed to fix Exceptions regarding incorrect types,
but it turned into this change.