Attempt to recover recording even if URL is present
We only attempted to recover recordings without a URL value because this
could mean that someone could keep re-uploading their recordings, but
since this command is only for supporters and myself, no need to do
this.

If I were to guard for that, this check should be done in a different
place.