Improve logging and lock use in CombinedAudioRecorderHandler
Attempted to commit these changes in pieces but they were too many
related changes.
* Convert this into a single use class, so we could make vars into vals
  which helped us remove the `init` clause and turn .? -> .
* Add warn logs before lock operations to help us see where the
  deadlocks may be happening.
* Re-arranged the order of the locks, could potentially have issues but
  will test and see what happens.
* In `saveRecording`, a FileOutputStream was opened and kept open until
  many operations happened. Instead, we open it do our queue -> mp3
  transform then close it. I bet that was causing some issues.

Overall, it seems like we have many subscribers and we use locks to
control the flow. This goes against RxJava nature and this makes it
harder to work with. So my next iteration will be to use `.blockingGet`
and get that `QueueFile` instance and work against it. It's a miracle
this works at all :sweat_smile: