Update pawa#recoverRecording to return a RecordingReply
This change trickles down to many parts of the code, but this is
necessary to be able to explain why a recovery failed. This is a quality
of life improvement because this is the only stopgap for the failures
that happen in production that I haven't been able to identify.

This will also allow us to eventually enable this for others to use
without being a security issue.