Add coverage maven profile
To generate coverage reports, we need the additonal `argLine` property
to be set, but we also need a specific property because of the mockk
test library. So to keep things separate, we've created a `coverage`
profile to define the JaCoCo plugin and configure surefire the way it
expects it to. win-win