Change .queue() calls to .await() to take advantage of coroutines
Using `.queue()` is the old way of doing things, and requires callbacks
for handling success and failure states. With `CoroutineEventListener`,
I can use `.await()` and when the code is running another coroutine can
execute while I wait for result.

Now if there's a failure in the interaction, it'll fail on the spot
instead of continuing.